Create a separate CommentTest
